First version of GenCustomers2.py

+++ GenCustomers2.py
@@ -3,7 +3,6 @@
 ## TO DO:
 # All Customers Page
 # All Pools Page
-# Check Andre email
 
 ### TABLE OF CONTENTS
 """
@@ -60,10 +59,10 @@ if True:
     MAX_WEEKS = 52
     DUMP_FOLDER_PATH = "../BNRPools/print_populated_pools/Dump Folder"
     CUSTOMERS_POOLS_XLSX = "template.xlsx"
-    C_ID_COL = 2
-    C_ID_START_ROW = 2
+    C_ID_COL = 1
+    C_ID_START_ROW = 1
     C_ID_END_ROW = 1000
-    P_ID_START_COL = 3
+    P_ID_START_COL = 2
     P_ID_END_COL = 1000
 
 ### Pseudocode
@@ -119,13 +118,26 @@ if True:
     
     
     # iterate through the CustomerID column in the xlsx
-    for row in range(C_ID_START_ROW, C_ID_END_ROW):
+    for count in range(C_ID_START_ROW, C_ID_END_ROW):
+        row = 2*count+1
         CustomerID = worksheet.cell(row=row, column=C_ID_COL).value
         if CustomerID is None:
             break
+        CustomerID = CustomerID.strip()
+        CustomerName = worksheet.cell(row=row+1, column=C_ID_COL).value
         print(f"\n CustomerID: {CustomerID}")
+        print(f"\n CustomerName: {CustomerName}")
         # for each CustomerID
         # create a CustomerID folder in the Customers folder
+        cidFolder = output_dir + "/"+CustomerID
+        # check to see if this CustomerID was already caught
+        tempcount=0
+        extraString = ""
+        while os.path.exists(cidFolder + extraString):
+            tempcount+=1
+            extraString = "AND"+str(tempcount)
+            # temporary solution
+        CustomerID=CustomerID+extraString
         cidFolder = output_dir + "/"+CustomerID
         os.makedirs(cidFolder, exist_ok=True)
         # print(f" Created cidFolder:{cidFolder}") # debug
@@ -133,14 +145,16 @@ if True:
         # create a CustomerID.html file in that CustomerID folder
         cid_html_name = CustomerID+".html"
         cid_html_path = cidFolder + "/" + cid_html_name
-        cid_html_content= f"<!DOCTYPE html>\n<html>\n<head>\n<title>{CustomerID}</title>\n</head>\n<body>\n<h2>{CustomerID}</h2>\n<ul>\n"
+        cid_html_content= f"<!DOCTYPE html>\n<html>\n<head>\n<title>{CustomerID}</title>\n</head>\n<body>\n<h2>{CustomerID}</h2>\n<h3>{CustomerName}</h3>\n<ul>\n"
         
         # iterate through the row of the CustomerID in the xlsx
         for col in range(P_ID_START_COL, P_ID_END_COL):
             PoolID = worksheet.cell(row=row, column=col).value
             if PoolID is None:
                 break
-            print(f"   PoolID: {PoolID}")
+            PoolID = PoolID.strip()
+            PoolName = worksheet.cell(row=row+1, column=col).value
+            print(f"   {PoolID}: {PoolName}")
             # for each PoolID
             # create a PoolID folder in the CustomerID folder
             pidFolder = cidFolder+"/"+PoolID
@@ -150,40 +164,40 @@ if True:
             # create a PoolID.html file in that PoolID folder
             pid_html_name = PoolID+".html"
             pid_html_path = pidFolder + "/" + pid_html_name
-            pid_html_content= f"<!DOCTYPE html>\n<html>\n<head>\n<title>{PoolID}</title>\n</head>\n<body>\n<h2>{PoolID}</h2>\n<ul>\n"
-            
+            pid_html_content= f"<!DOCTYPE html>\n<html>\n<head>\n<title>{PoolID}</title>\n</head>\n<body>\n<h2>{PoolID}</h2>\n<h3>{PoolName}</h3>\n<ul>\n"
             # iterate through the PoolID folder in the dumpfile
             Dump_Pool_ID_path = DUMP_FOLDER_PATH+"/"+str(PoolID)
-            fnames = os.listdir(Dump_Pool_ID_path)
-            for file in range(0,MAX_WEEKS):
-                try:
-                    fileName = os.path.basename(fnames[file])
-                    # print(f"\t{fileName}") #debug
-                    # import and rename the pdf into the PoolID folder
-                    # Source path
-                    src_path = Dump_Pool_ID_path + "/" + fileName
-
-                    # Destination path
-                    dest_path = pidFolder+ "/" + fileName
-                    
-                    # Copy the file
+            if os.path.exists(Dump_Pool_ID_path):
+                fnames = os.listdir(Dump_Pool_ID_path)
+                for file in range(0,MAX_WEEKS):
                     try:
-                        # print(f"\t Copying from\t {src_path}\n\t to\t {dest_path}") # debug
-                        shutil.copy2(src_path, dest_path)
-                        # print(f"\t Copy Success: {fileName}") # debug
+                        fileName = os.path.basename(fnames[file])
+                        # print(f"\t{fileName}") #debug
+                        # import and rename the pdf into the PoolID folder
+                        # Source path
+                        src_path = Dump_Pool_ID_path + "/" + fileName
+
+                        # Destination path
+                        dest_path = pidFolder+ "/" + fileName
+                        
+                        # Copy the file
+                        try:
+                            # print(f"\t Copying from\t {src_path}\n\t to\t {dest_path}") # debug
+                            shutil.copy2(src_path, dest_path)
+                            # print(f"\t Copy Success: {fileName}") # debug
+                        except Exception as e:
+                            print(f"\t Copy Failed!:{e}")
+                        
+                        # add a link to the PoolID.html that opens PoolDate.pdf
+                        pid_html_content += f'    <li><a target="_blank" href="{fileName}">{fileName}</a></li>\n'
+                        
+                    except IndexError as e:
+                        # print(f"{e}") # debug
+                        break
                     except Exception as e:
-                        print(f"\t Copy Failed!:{e}")
-                    
-                    # add a link to the PoolID.html that opens PoolDate.pdf
-                    pid_html_content += f'    <li><a target="_blank" href="{fileName}">{fileName}</a></li>\n'
-                    
-                except IndexError as e:
-                    # print(f"{e}") # debug
-                    break
-                except Exception as e:
-                    print(style.RED + f"!--ERROR:{e}" + style.RESET)
-                    break
-            print(f"\t {file} pdfs transferred")
+                        print(style.RED + f"!--ERROR:{e}" + style.RESET)
+                        break
+                print(f"\t {file} pdfs transferred")
             
             # save PoolID.html
             pid_html_content += "</ul>\n</body>\n</html>"
@@ -202,7 +216,7 @@ if True:
         # print(f" {cid_html_name} has been generated.") # debug
         
         # add a link to AllCustomers.html that leads to CustomerID.html
-        all_cids_html_content += f'    <li><a target="bottom" href="{CustomerID}/{cid_html_name}">{CustomerID}</a></li>\n'
+        all_cids_html_content += f'    <li><a target="bottom" href="{CustomerID}/{cid_html_name}">{CustomerID}</a> {CustomerName}</li>\n'
 
 
     # save AllCustomers.html in Customers Folder
